Improved form validations - Forgot password form
